v1.0.17

+++ src/Service/ClearDeletedIndexes.php
@@ -2,6 +2,7 @@
 
 namespace IntegrationHelper\IntegrationVersionLaravelServer\Service;
 
+use Illuminate\Support\Facades\Cache;
 use IntegrationHelper\IntegrationVersion\Context;
 use IntegrationHelper\IntegrationVersion\IntegrationVersionManagerInterface;
 use IntegrationHelper\IntegrationVersion\IntegrationVersionItemManagerInterface;
@@ -23,27 +24,30 @@ class ClearDeletedIndexes
          */
         $idsToDelete = [];
         $sources = [];
+        $locker = Cache::lock('integration-version-clear-deleted', 3600);
 
-        foreach (
-            $this->integrationVersionItemManager->getItemsWithDeletedStatus() as
-            $items
-        ) {
-            foreach ($items as $item) {
-                $idsToDelete[] = $item->getIdValue();
-                $sources[$item->getParentId()] = $item->getParentId();
+        if($locker->get() === false) {
+            foreach (
+                $this->integrationVersionItemManager->getItemsWithDeletedStatus() as
+                $items
+            ) {
+                foreach ($items as $item) {
+                    $idsToDelete[] = $item->getIdValue();
+                    $sources[$item->getParentId()] = $item->getParentId();
+                }
             }
-        }
-        $hashDateTime = Context::getInstance()->getDateTime()->getNow();
+            $hashDateTime = Context::getInstance()->getDateTime()->getNow();
 
-        if($idsToDelete) {
-            foreach (array_chunk($idsToDelete, 1000) as $chunk) {
-                $this->integrationVersionItemManager->delete($chunk);
-            }
-            if($sources) {
-                foreach ($sources as $sourceId) {
-                    $inventoryVersion = $this->integrationVersionRepository->getItemById($sourceId);
-                    $hash = Context::getInstance()->getHashGenerator()->generate($inventoryVersion->getSource());
-                    $this->integrationVersionManager->saveNewHash($inventoryVersion->getSource(), $hash, $hashDateTime);
+            if($idsToDelete) {
+                foreach (array_chunk($idsToDelete, 1000) as $chunk) {
+                    $this->integrationVersionItemManager->delete($chunk);
+                }
+                if($sources) {
+                    foreach ($sources as $sourceId) {
+                        $inventoryVersion = $this->integrationVersionRepository->getItemById($sourceId);
+                        $hash = Context::getInstance()->getHashGenerator()->generate($inventoryVersion->getSource());
+                        $this->integrationVersionManager->saveNewHash($inventoryVersion->getSource(), $hash, $hashDateTime);
+                    }
                 }
             }
         }
